requirements.txt
discord.py==2.1.1
peewee==3.16.2
python-dotenv==1.0.0
Requests==2.31.0
